Obtain command-line arguments before PSP is trashed
If SPAWNO is utilized, it trashes the PSP, which is where command-line
arguments are stored.  We execute freeram.com at startup to get a real,
working idea of exactly how much RAM is available for launching
programs.  To prevent losing our original command-line options after
this happens, we grab what we need and store it in program structures
before the PSP is trashed.